+++ src/lib/descriptor/mod_test.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-use crate::types::{ExtendOptions, InstallCrate};
+use crate::types::{ExtendOptions, InstallCrate, ScriptValue};
 
 #[test]
 fn merge_env_both_empty() {
@@ -724,7 +724,7 @@ fn run_load_script_no_load_script() {
 #[test]
 fn run_load_script_valid_load_script() {
     let mut config = ConfigSection::new();
-    config.load_script = Some(vec!["exit 0".to_string()]);
+    config.load_script = Some(ScriptValue::Text(vec!["exit 0".to_string()]));
 
     let mut external_config = ExternalConfig::new();
     external_config.config = Some(config);
@@ -737,7 +737,7 @@ fn run_load_script_valid_load_script() {
 #[should_panic]
 fn run_load_script_invalid_load_script() {
     let mut config = ConfigSection::new();
-    config.load_script = Some(vec!["exit 1".to_string()]);
+    config.load_script = Some(ScriptValue::Text(vec!["exit 1".to_string()]));
 
     let mut external_config = ExternalConfig::new();
     external_config.config = Some(config);
@@ -753,10 +753,10 @@ fn run_load_script_valid_load_script_duckscript() {
     ));
 
     let mut config = ConfigSection::new();
-    config.load_script = Some(vec![r#"#!@duckscript
+    config.load_script = Some(ScriptValue::Text(vec![r#"#!@duckscript
     set_env run_load_script_valid_load_script_duckscript true
     "#
-    .to_string()]);
+    .to_string()]));
 
     let mut external_config = ExternalConfig::new();
     external_config.config = Some(config);